+++ lib/modelrom.cc
@@ -17,7 +17,7 @@ ModelRom::Segment::operator <(uint32_t rhs_address) const {
 }
 
 bool
-ModelRom::Segment::contains(uint32_t address, uint8_t size) const {
+ModelRom::Segment::contains(uint32_t address, uint16_t size) const {
   return (this->address <= address)
       && ((this->address+content.size()) >= (address+size));
 }
@@ -80,7 +80,7 @@ ModelRom::write(uint32_t address, const QByteArray &data) {
 
 
 bool
-ModelRom::read(uint32_t address, uint8_t length, QByteArray &data) const {
+ModelRom::read(uint32_t address, uint16_t length, QByteArray &data) const {
   auto next = std::lower_bound(begin(), end(), address);
 
   if (begin() == next) {
